requirements.txt
fastapi
uvicorn
boto3
python-multipart
openai
Pillow
